Separating unit tests
Separating unit tests for the InteractsWithPages and MakesHttpRequests
class

On MakesHttpRequestsTest, the methods:
- prepareUrlForRequest
- seeStatusCode
- assertResponseOk
- assertResponseStatus

belong to MakesHttpRequests class and the rest of the methods belong to
InteractsWithPages

Changes to be committed:
    copied:     tests/Unit/MakesHttpRequestsTest.php -> tests/Unit/InteractsWithPagesTest.php
    modified:   tests/Unit/MakesHttpRequestsTest.php
